add member role, process swagger issue
