+++ Backend/src/main/java/com/fintech/moang/contorller/AuthController.java
@@ -6,7 +6,9 @@ import com.fintech.moang.domain.AuthRefreshToken.AuthRefreshToken;
 import com.fintech.moang.domain.AuthRefreshToken.AuthRefreshTokenRepository;
 import com.fintech.moang.domain.member.Member;
 import com.fintech.moang.domain.member.MemberRepository;
+import com.fintech.moang.dto.request.MemberJoinPostReqDto;
 import com.fintech.moang.dto.request.MemberLoginPostReqDto;
+import com.fintech.moang.dto.response.MemberJoinPostResDto;
 import com.fintech.moang.dto.response.MemberLoginPostResDto;
 import com.fintech.moang.service.MemberService;
 import io.swagger.annotations.*;
@@ -30,6 +32,23 @@ public class AuthController {
     private final PasswordEncoder passwordEncoder;
 
     private final AuthRefreshTokenRepository authRefreshTokenRepository;
+
+    @PostMapping()
+    @ApiOperation(value="회원 가입", notes = "아이디와 패스워드를 통해 회원가입 한다")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "성공"),
+            @ApiResponse(code = 401, message = "가입 실패"),
+            @ApiResponse(code = 404, message = "반환값 없음"),
+            @ApiResponse(code = 500, message = "서버 오류"),
+    })
+    public ResponseEntity<MemberJoinPostResDto> register(@RequestBody @ApiParam(value="회원가입 정보", required = true) MemberJoinPostReqDto joinInfo) {
+        if(memberService.createMember(joinInfo)) {
+            // 가입 성공시 성공 리턴
+            return ResponseEntity.status(200).body(MemberJoinPostResDto.of(200, "Success", true));
+        }
+        // 가입 실패시 실패 리턴
+        return ResponseEntity.status(401).body(MemberJoinPostResDto.of(401, "Failure", false));
+    }
 
     @PostMapping("/login")
     @ApiOperation(value = "로그인", notes = "아이디와 패스워드를 통해 로그인 한다")
